ModelGenerator-tensorflow/requirements.txt
keras
tensorflow
# tensorflow-gpu # Alternatively if available
codecov
pytest
pytest-cov
matplotlib
scipy
pillow
tqdm